phase1086-lint-internal-linkage-and-redundant-casting(engine+samples)
Warning-class wave 1 of the Sprint-1 leftovers: misc-use-internal-
linkage (10 sites) + readability-redundant-casting (1 site) cleaned.

Genuinely-internal functions gain static (D3D12ShaderCompile sm5/sm6
compilers — called in-TU only; two hello_engine inline helpers).
The Wasapi stream-id cast was redundant (already uint32) and is gone.

The check is BLIND to the project's local-extern-declaration pattern,
and its auto-fix broke the link THREE separate ways before review:
Vulkan create_device (consumed by VulkanDeviceFactory.cpp),
ft_backend::load/rasterize_glyph + msdfgen_backend::to_msdf_multi
(consumed by Font.cpp), and the two net TU-anchor functions (consumed
by test_net_qos/test_net_reconciler via extern decls — caught by
lld undefined-symbol, restored). All six carry NOLINT with the exact
cross-TU consumer named. Lesson reinforced: single-rule --fix output
is REVIEWED against link reality, never trusted.

Probe re-run over all offender TUs: 0 diagnostics. 263/263 ctest.
Promotion of both rules lands with the wave-2/3 close-out.

+++ engine/world/audio/src/WasapiAudioBackend.cpp
@@ -296,7 +296,7 @@ public:
                 "create_stream: channels / sample_rate must be > 0"));
         }
         std::scoped_lock guard { state_mu_ };
-        const auto id = static_cast<std::uint32_t>(next_stream_id_++);
+        const auto id = next_stream_id_++;
         StreamRec rec;
         rec.channels = channels;
         rec.sample_rate = sample_rate;
